+++ src/views/Sections/Character.vue
