+++ src/views/Sections/Declarations.vue
