+++ src/views/Sections/Diversity.vue
